% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lira_on_images.R
\name{detect_sources_LIRA}
\alias{detect_sources_LIRA}
\title{Run LIRA to detect sources using the configuration from config.yaml}
\usage{
detect_sources_LIRA()
}
\description{
Run LIRA to detect sources using the configuration from config.yaml
}
