% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_p_computation.R
\name{generate.distribution_xi}
\alias{generate.distribution_xi}
\title{Compute the xi for all the iterations in each image}
\usage{
generate.distribution_xi(images, mask_file, null_file, n_iter)
}
\description{
Compute the xi for all the iterations in each image
}
