% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_p_computation.R
\name{save_distribution_xi_t}
\alias{save_distribution_xi_t}
\title{Take the distibutions of xi, save them as pdf/mat files, and compute the upper bound on p value}
\usage{
save_distribution_xi_t(xi.distribution, region.name, out.dir)
}
\description{
Take the distibutions of xi, save them as pdf/mat files, and compute the upper bound on p value
}
