% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_p_computation.R
\name{save_distributions_xi}
\alias{save_distributions_xi}
\title{Take the distibutions of xi, save them as pdf/mat files, and compute the unpper bound on p value}
\usage{
save_distributions_xi(distributions, region_name, out_dir)
}
\description{
Take the distibutions of xi, save them as pdf/mat files, and compute the unpper bound on p value
}
