% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lira_on_images.R
\name{detect_sources_LIRA}
\alias{detect_sources_LIRA}
\title{Run LIRA to detect sources using the configuration from config.yaml
Following are the required parameters in config.yaml
obs_file: This is the X-ray observation file with square dimensions
psf_file: PSF file. It is assumed that this file represents the psf at all points in the image
null_file: The null model against which the observation will be compared against
replica_im_template: template name of the images replicated from the null/baseline model.
                     For example, if the replicated images are named sim_64x64_0.5_{0...n}.fits, 
                     the template will be sim_64x64_0.5_%s.fits
n_replicas: The number of replicated images. The count should start from 0}
\usage{
detect_sources_LIRA()
}
\description{
Optional parameters
max_iter: Total number of draws from the LIRA posterior. Defaults to 2000
burn_in: Total number of draws to be ignored from the beginning. Defaults to 1000
alpha.init: initial values of the smoothing parameters. Defaults to [0.3,0.4,0.5,0.6,0.7,0.8]
thin: Number of {thin}th draws to store. Defaults to 1
output_dir: Output directory to store all the outputs. Defaults to LIRA_outputs
n_cores: LIRA will be run simultaneously on n_cores each processing a different image.
         Defaults to all the available cores.
}
